+++ test.py
@@ -20,17 +20,24 @@ from __future__ import print_function
 
 import sys
 
-if __name__ == '__main__':
+def testmod(mod):
     import doctest
-    import NumRep
-    err, test = doctest.testmod(m=NumRep,verbose=True)
+    mod = __import__(mod)
+    err, test = doctest.testmod(m=mod,verbose=True)
     if err > 0:
-        sys.exit(1)
+        return False
     else:
-        print()
-        print("All tests passed!")
-        sys.exit(0)
+        return True
 
-else:
-    sys.stderr.write('Can be only run as a single program\n')
-    sys.exit(1)
+if __name__ == '__main__':
+    import os
+    first = testmod('NumRep')
+    if not first:
+        sys.exit(1)
+    for mod in os.listdir('NumRep'):
+        if mod.startswith('__') or mod.endswith('.pyc'):
+            continue
+        module = 'NumRep.'+mod.replace('.py','')
+        test = testmod(module)
+        if not test:
+            sys.exit(1)