+++ CHANGELOG.rst
@@ -1,6 +1,6 @@
 0.1
  * Now can compare and do arithmetic operations on NumRep objects
- * Changed NumRep to a package
+ * Made NumRep a package
 
 0.0.1.1
  * Added variabes for all() functions in the module